New module named crud.js Creat a function to add an new artist to collection by id
