+++ modules/render.js
@@ -1,9 +1,4 @@
-import {
-  saveListToLS,
-  ARTISTS_KEY,
-  // SEARCH_KEY,
-  loadSearchListFromLS,
-} from "./localStorage.js";
+import { addArtistToCollection } from "./crud.js";
 
 // DOM references
 const searchResultUlElem = document.querySelector(".search-result-list");
@@ -12,9 +7,6 @@ const searchResultUlElem = document.querySelector(".search-result-list");
 // let searchResultList = [];
 // searchResultList = loadSearchListFromLS(SEARCH_KEY);
 // renderSearchResults(searchResultList);
-
-let artistsList = [];
-artistsList = loadSearchListFromLS(ARTISTS_KEY);
 
 export function renderSearchResults(array) {
   array.forEach((a) => {
@@ -44,22 +36,7 @@ export function renderSearchResults(array) {
 
     // Eventlistener
     addBtnElem.addEventListener("click", () => {
-      const id = a.id;
-
-      // Check if the artist already in the collection
-      const artistAlreadyListed = artistsList.find(
-        (listedArtist) => listedArtist.id === id
-      );
-
-      if (!artistAlreadyListed) {
-        // Add the artist object to the array and save to local storage
-        artistsList.push(a);
-        saveListToLS(ARTISTS_KEY, artistsList);
-      } else {
-        alert("Artist is already in your collection"); // IF TIME; FIX A POPUP OR MODAL INSTEAD
-      }
-
-      // Render artist-collection function here
+      addArtistToCollection(a.id);
     });
   });
 }
